lg-quick-start/pyproject.toml
[project]
name = "lg-quick-start"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.5",
    "langchain-openai>=0.2.4",
    "langgraph>=0.2.39",
    "python-dotenv>=1.0.1",
    "ipython>=8.29.0",
    "tavily-python>=0.5.0",
    "langchain-community>=0.3.3",
]
